Changed to UMD format for cross compatibilty (node/browser)
