+++ src/event-kit.coffee
@@ -2,6 +2,4 @@ import Emitter from './emitter.coffee'
 import Disposable from './disposable.coffee'
 import CompositeDisposable from './composite-disposable.coffee'
 
-window.Emitter = Emitter
-window.Disposable = Disposable
-window.CompositeDisposable = CompositeDisposable
+export {Emitter, Disposable, CompositeDisposable}
